fastapi_app/backend/requirements.txt
fastapi
uvicorn
python-multipart
requests
python-dotenv
ultralytics
pillow

